namespaceの修正とダメージUsecaseのテスト追加
