+++ Assets/Plugins/BlackSmith.Domain/Domain/Character/Battle/AttackValue.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 1c244466f09275d44a4015b89fb08054
