+++ Assets/Plugins/BlackSmith.Domain/Domain/Character/Battle/DamageValue.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 00af58d1abad89c4299ad9307228845f
