+++ Assets/Plugins/BlackSmith.Domain/Domain/Character/Battle/DefenseValue.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 885f5c4043e4fdf4680e7df52c4b1627
