+++ Assets/Plugins/BlackSmith.Domain/Domain/Character/Battle/Equipment.cs
