+++ Assets/Plugins/BlackSmith.Domain/Domain/Character/Battle/Equipment.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 506cccb86541f84428c21d32128dc05f
