+++ Assets/Plugins/BlackSmith.Domain/Domain/Character/Battle/HealthPoint.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: afe4204fac5a07842b90fac7d4f9ae26
