+++ Assets/Plugins/BlackSmith.Domain/Domain/Character/Battle/IBattleCharacter.cs
@@ -1,5 +1,4 @@
-﻿using BlackSmith.Domain.CharacterObject;
-using BlackSmith.Domain.CharacterObject.Interface;
+﻿using BlackSmith.Domain.CharacterObject.Interface;
 
 namespace BlackSmith.Domain.Character.Battle
 {
