+++ Assets/Plugins/BlackSmith.Domain/Domain/Character/Battle/ITakeDamageable.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using BlackSmith.Domain.Character.Battle;
+using System;
 
 namespace BlackSmith.Domain.CharacterObject.Interface
 {
