+++ Assets/Plugins/BlackSmith.Domain/Domain/Character/Battle/PlayerBattleEntity.cs
@@ -1,5 +1,4 @@
-﻿using BlackSmith.Domain.CharacterObject;
-using BlackSmith.Domain.CharacterObject.Interface;
+﻿using BlackSmith.Domain.CharacterObject.Interface;
 using BlackSmith.Domain.Item.Equipment;
 using System;
