+++ Assets/Plugins/BlackSmith.Domain/Domain/Character/Player/Event/PlayerHealthChangedEvent.cs
@@ -1,4 +1,4 @@
-﻿using BlackSmith.Domain.CharacterObject;
+﻿using BlackSmith.Domain.Character.Battle;
 
 namespace BlackSmith.Domain.Character.Player.Event
 {
