--- Assets/Plugins/BlackSmith.Domain/Domain/CharacterObject.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 749ec1c560b535f41abc76a1f7c2dbbc
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
