--- Assets/Plugins/BlackSmith.Domain/Domain/CharacterObject/AttackValue.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 06ed95ad25a208e4a9f575926f9a6ab2
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
