--- Assets/Plugins/BlackSmith.Domain/Domain/CharacterObject/DamageValue.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: e5e345ff6c764f24cab9878c11efd82e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
