--- Assets/Plugins/BlackSmith.Domain/Domain/CharacterObject/DefenseValue.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: a2dd76d963027404d80c439f6e88727b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
