--- Assets/Plugins/BlackSmith.Domain/Domain/CharacterObject/Equipment.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: eeee6ab448f865e4a994ab12b0863634
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
