--- Assets/Plugins/BlackSmith.Domain/Domain/CharacterObject/HealthPoint.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 4d0dd9cec77741145863cfbeb01cae0a
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
