--- Assets/Plugins/BlackSmith.Domain/Domain/CharacterObject/Interface.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 27fb098bc7b0eac48bcb227930030369
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
