+++ Assets/Plugins/BlackSmith.Domain/Tests/Domain/Battle/PlayerBattleEntityTest.cs
@@ -1,7 +1,6 @@
 ﻿using BlackSmith.Domain.Character;
 using BlackSmith.Domain.Character.Battle;
 using BlackSmith.Domain.Character.Player;
-using BlackSmith.Domain.CharacterObject;
 using NUnit.Framework;
 using System;
 using System.Collections;
